perf(ui): Reduce background blur intensity and clean up theme color code
- Lower background blur radius from 50 to 15 improves rendering performance
- Remove unused old rgbToHslFloat implementation and dead helper functions
- Migrate all color conversion calls to use standard shared utility functions
- Update palette generation call signatures to match refactored helpers
- Clean up unused imports from theme extractor module
